requirements.txt
numpy>=1.18.2
matplotlib>=3.1.1
tqdm>=4.45.0
nltk>=3.4.5
Pillow>=7.1.1
pycocotools>=2.0
seaborn>=0.9.0
torch>=1.4.0
torchvision>=0.5.0
